+++ src/lib.rs
@@ -30,7 +30,7 @@ pub extern crate bitcoin;
 extern crate bitcoin_hashes as just_imported_for_the_macros;
 extern crate slip21;
 pub extern crate secp256k1_zkp;
-#[cfg(feature = "serde")] #[macro_use] extern crate serde;
+#[cfg(feature = "serde")] #[macro_use] extern crate actual_serde as serde;
 #[cfg(all(test, feature = "serde"))] extern crate serde_test;
 
 #[cfg(test)] extern crate rand;